Crear PLaylist listo

+++ Entrega 3/Clases/SongClass.cs
@@ -17,7 +17,7 @@ namespace Entrega_3.Clases
         private string gender;
         private string publicationYear; //Tiene que ser DAte time
         private string title;
-        private int duration;
+        private double duration;
         private int memory;
         private string study;
         private string keyword;
@@ -36,7 +36,7 @@ namespace Entrega_3.Clases
         public string Gender { get => gender; set => gender = value; }
         public string PublicationYear { get => publicationYear; set => publicationYear = value; }
         public string Title { get => title; set => title = value; }
-        public int Duration { get => duration; set => duration = value; }
+        public double Duration { get => duration; set => duration = value; }
         public int Memory { get => memory; set => memory = value; }
         public string Study { get => study; set => study = value; }
         public string Keyword { get => keyword; set => keyword = value; }
@@ -48,7 +48,7 @@ namespace Entrega_3.Clases
         public int Likes { get => likes; set => likes = value; }
 
         public SongClass(string gender, string publicationYear,
-                     string title, int duration, int memory, string study, string keyword,
+                     string title, double duration, int memory, string study, string keyword,
                      string composer, Singer singer, Album album, string format, int evaluation, int resolution, string url, int nReproduction, int likes)
         {
             this.Gender = gender; // listo
